requirements.txt
google-generativeai==0.8.5
Jinja2==3.1.6
numpy==2.3.0
pandas==2.3.3
pillow==11.2.1
PyMuPDF==1.26.1
python-docx==1.1.2
RapidFuzz==3.13.0
scikit-learn==1.8.0
sentence-transformers==3.0.1
spacy==3.8.7
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
streamlit==1.45.1